camera: Create span using unsigned length
A recent CL https://crrev.com/c/4219636 checks that span sizes are
unsigned.

This CL changes the type of the length of a span to be unsigned to meet
the requirements. It is required for libchrome uprev r1103195.

BUG=None
TEST=emerge-${BOARD} cros-camera-libs libchrome of new and old.
TEST=CQ

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4252160
Tested-by: Cici Ruan <cuicuiruan@google.com>
Auto-Submit: Cici Ruan <cuicuiruan@google.com>
Commit-Queue: Cici Ruan <cuicuiruan@google.com>

+++ camera/gpu/gles/screen_space_rect.cc
@@ -48,7 +48,7 @@ Buffer CreateVertexBuffer(float x, float y, float width, float height) {
   const uint8_t* buffer_as_bytes =
       reinterpret_cast<const uint8_t*>(raw_buffer.data());
   Buffer buffer(kBufferSizeBytes, GL_STATIC_DRAW);
-  buffer.Set({buffer_as_bytes, kBufferSizeBytes});
+  buffer.Set({buffer_as_bytes, static_cast<size_t>(kBufferSizeBytes)});
   return buffer;
 }
